+++ src/client/components/CoreUI.js
@@ -1,6 +1,6 @@
 import { css } from '@firebolt-dev/css'
 import { useEffect, useMemo, useRef, useState } from 'react'
-import { LoaderIcon, MessageSquareTextIcon } from 'lucide-react'
+import { LoaderIcon, MessageSquareTextIcon, RefreshCwIcon } from 'lucide-react'
 import moment from 'moment'
 
 // import { CodeEditor } from './CodeEditor'
@@ -90,7 +90,7 @@ export function CoreUI({ world }) {
       `}
     >
       {disconnected && <Disconnected />}
-      <Reticle world={world} />
+      {!ui.reticleSuppressors && <Reticle world={world} />}
       {<Toast world={world} />}
       {ready && <ActionsBlock world={world} />}
       {ready && <Sidebar world={world} ui={ui} />}
@@ -635,27 +635,54 @@ function Disconnected() {
   //   }
   // }, [])
   return (
-    <div
-      css={css`
-        position: fixed;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-        backdrop-filter: grayscale(100%);
-        pointer-events: none;
-        z-index: 9999;
-        animation: fadeIn 3s forwards;
-        @keyframes fadeIn {
-          from {
-            opacity: 0;
+    <>
+      <div
+        css={css`
+          position: fixed;
+          top: 0;
+          left: 0;
+          width: 100%;
+          height: 100%;
+          backdrop-filter: grayscale(100%);
+          pointer-events: none;
+          z-index: 9999;
+          animation: fadeIn 3s forwards;
+          @keyframes fadeIn {
+            from {
+              opacity: 0;
+            }
+            to {
+              opacity: 1;
+            }
           }
-          to {
-            opacity: 1;
+        `}
+      />
+      <div
+        css={css`
+          pointer-events: auto;
+          position: absolute;
+          top: 50%;
+          left: 50%;
+          background: rgba(11, 10, 21, 0.85);
+          border: 0.0625rem solid #2a2b39;
+          backdrop-filter: blur(5px);
+          border-radius: 1rem;
+          height: 2.75rem;
+          padding: 0 1rem;
+          transform: translate(-50%, -50%);
+          display: flex;
+          align-items: center;
+          cursor: pointer;
+          > span {
+            margin-left: 0.4rem;
           }
-        }
-      `}
-    />
+        `}
+        onClick={() => window.location.reload()}
+      >
+        <RefreshCwIcon size='1.1rem' />
+        <span>Reconnect</span>
+      </div>
+    </>
   )
 }