+++ src/core/systems/ClientControls.js
@@ -32,9 +32,11 @@ const controlTypes = {
   screen: createScreen,
   camera: createCamera,
   xrLeftStick: createVector,
+  xrLeftTrigger: createButton,
   xrLeftBtn1: createButton,
   xrLeftBtn2: createButton,
   xrRightStick: createVector,
+  xrRightTrigger: createButton,
   xrRightBtn1: createButton,
   xrRightBtn2: createButton,
   touchA: createButton,
@@ -90,6 +92,19 @@ export class ClientControls extends System {
               control.entries.xrLeftStick.value.z = src.gamepad.axes[3]
               if (control.entries.xrLeftStick.capture) break
             }
+            if (control.entries.xrLeftTrigger) {
+              const button = control.entries.xrLeftTrigger
+              const down = src.gamepad.buttons[0].pressed
+              if (down && !button.down) {
+                button.pressed = true
+                button.onPress?.()
+              }
+              if (!down && button.down) {
+                button.released = true
+                button.onRelease?.()
+              }
+              button.down = down
+            }
             if (control.entries.xrLeftBtn1) {
               const button = control.entries.xrLeftBtn1
               const down = src.gamepad.buttons[4].pressed
@@ -125,6 +140,19 @@ export class ClientControls extends System {
               control.entries.xrRightStick.value.x = src.gamepad.axes[2]
               control.entries.xrRightStick.value.z = src.gamepad.axes[3]
               if (control.entries.xrRightStick.capture) break
+            }
+            if (control.entries.xrRightTrigger) {
+              const button = control.entries.xrRightTrigger
+              const down = src.gamepad.buttons[0].pressed
+              if (down && !button.down) {
+                button.pressed = true
+                button.onPress?.()
+              }
+              if (!down && button.down) {
+                button.released = true
+                button.onRelease?.()
+              }
+              button.down = down
             }
             if (control.entries.xrRightBtn1) {
               const button = control.entries.xrRightBtn1
@@ -212,17 +240,29 @@ export class ClientControls extends System {
     this.viewport.addEventListener('wheel', this.onScroll, { passive: false })
     document.body.addEventListener('contextmenu', this.onContextMenu)
     window.addEventListener('resize', this.onResize)
+    window.addEventListener('focus', this.onFocus)
     window.addEventListener('blur', this.onBlur)
   }
 
   bind(options = {}) {
     const self = this
     const entries = {}
+    let reticleSupressor
     const control = {
       options,
       entries,
       actions: null,
       api: {
+        hideReticle(value = true) {
+          if (reticleSupressor && value) return
+          if (!reticleSupressor && !value) return
+          if (reticleSupressor) {
+            reticleSupressor?.()
+            reticleSupressor = null
+          } else {
+            reticleSupressor = self.world.ui.suppressReticle()
+          }
+        },
         setActions(value) {
           if (value !== null && !Array.isArray(value)) {
             throw new Error('[control] actions must be null or array')
@@ -236,6 +276,7 @@ export class ClientControls extends System {
           self.buildActions()
         },
         release: () => {
+          reticleSupressor?.()
           const idx = this.controls.indexOf(control)
           if (idx === -1) return
           this.controls.splice(idx, 1)
@@ -602,6 +643,10 @@ export class ClientControls extends System {
     this.screen.height = this.viewport.offsetHeight
   }
 
+  onFocus = () => {
+    this.releaseAllButtons()
+  }
+
   onBlur = () => {
     this.releaseAllButtons()
   }
@@ -629,6 +674,7 @@ export class ClientControls extends System {
     this.viewport.removeEventListener('wheel', this.onScroll, { passive: false })
     document.body.removeEventListener('contextmenu', this.onContextMenu)
     window.removeEventListener('resize', this.onResize)
+    window.removeEventListener('focus', this.onFocus)
     window.removeEventListener('blur', this.onBlur)
   }
 }